requirements.txt
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
ultralytics==8.3.160
paddlepaddle>=2.6.0
paddleocr>=2.7.0
python-dotenv
requests
face_recognition
cmake
dlib
fastapi==0.115.0
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
onvif-zeep==0.2.12
OPi.GPIO
tinytuya
